+++ CHANGELOG.md
@@ -9,6 +9,10 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
+### Changed
+
+- Changed resource display name in notifications.
+
 ## [1.2.0] - 2023-11-15
 
 ### Added